requirements.txt
astroid==2.1.0
certifi==2019.3.9
chardet==3.0.4
dataclasses==0.6
dj-database-url==0.5.0
Django==2.0.7
django-json-field==0.5.7
django-toolbox==0.2
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
djangotoolbox==0.9.2
django-crispy-forms
gunicorn==19.9.0
idna==2.8
isort==4.3.9
jsonfield==2.0.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.7
PyJWT==1.5.3
pylint==2.2.2
python-dateutil==2.8.0
pytz==2017.3
requests==2.21.0
six==1.12.0
sqlparse==0.2.4
typed-ast==1.3.1
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.1
django-cors-headers
django-rest-auth
django-allauth
